pyproject.toml
[project]
name = "meeting-rooms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2",
    "django-filter>=25.1",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
